pyproject.toml
[tool.poetry]
name = "async-api-team-9"
version = "0.1.0"
description = ""
authors = ["Mikhail Spiridonov <Muxa2793@mail.ru>"]
readme = "README.md"
packages = [{include = "async_api_team_9"}]

[tool.poetry.dependencies]
python = "^3.11"
elasticsearch = "7.8.0"
pre-commit = "^3.1.1"
orjson = "^3.8.7"
aiohttp = "^3.8.4"
redis = "^4.5.1"
uvicorn = {extras = ["standard"], version = "^0.21.0"}
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
gunicorn = "20.1.0"
fastapi = "^0.95.0"


[tool.poetry.group.tests.dependencies]
pytest-asyncio = "0.12.0"
pytest = "6.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
profile = "black"

[tool.black]
line-length = 120

[tool.flake8]
exclude = [".env", ".env.example", ".git", "__pycache__", "venv/", "env/", "*/tests/*", "*/migrations/*"]
max-line-length = 120
